+++ build.rs
@@ -1,22 +1,27 @@
 fn main() {
 	let source_files = vec![
-		"src/raw/package.rs",
-		"src/raw/cache.rs",
-		"src/raw/progress.rs",
-		"src/raw/config.rs",
-		"src/raw/util.rs",
-		"src/raw/records.rs",
-		"src/raw/depcache.rs",
-		"src/raw/pkgmanager.rs",
-		"src/raw/error.rs",
+		"src/cache.rs",
+		"src/progress.rs",
+		"src/config.rs",
+		"src/util.rs",
+		"src/records.rs",
+		"src/depcache.rs",
+		"src/pkgmanager.rs",
+		"src/error.rs",
+		"src/acquire.rs",
+		"src/iterators/package.rs",
+		"src/iterators/version.rs",
+		"src/iterators/dependency.rs",
+		"src/iterators/provider.rs",
+		"src/iterators/files.rs",
 	];
 
 
-	let mut cc_files = vec!["apt-pkg-c/progress.cc", "apt-pkg-c/error.cc"];
+	let mut cc_files = vec!["apt-pkg-c/error.cc"];
 
 	cxx_build::bridges(&source_files)
 		.files(&cc_files)
-		.flag_if_supported("-std=c++17")
+		.flag_if_supported("-std=c++14")
 		.compile("oma-apt");
 
 	println!("cargo:rustc-link-lib=apt-pkg");
@@ -35,6 +40,7 @@ fn main() {
 		"apt-pkg-c/pkgmanager.h",
 		"apt-pkg-c/error.h",
 		"apt-pkg-c/types.h",
+		"apt-pkg-c/acquire.h",
 	]);
 
 	for file in cc_files {